Removed check use logic from purgemod; updated API
Prior to this commit the purge_mod method was checking for use in an
environment and would fail if found.  This is not always desirable so I
removed this logic and added the `check_mod_use` method.  This method
validates if a module is in use, and if it is returns the environments
consuming the name/version pair.  The API purge_mod route was also
updated to use the `check_mod_use` method and will return a failure if
it is in use along with the environments it was found in.