Make redeploy durable
Prior to this commit redeployments that failed half-way through would
result in a inconsistent environment dir.  If this went unresolved the
next deploy that ran would force file-sync to commit an inconsistent
directory.  This commit implements a temp backup system that will revert
the previous version in the event of a failed checkout.